🔧 Implement JDA configuration & initialization
Implemented a JDA Configuration which looks for a token property in
the custom property file and attempts to instantiate a JDA builder
with it. This configuration also registers all ListenerAdapters
that are loaded into Spring.

Added a custom lookup for a scriptutils.properties in the cwd from
which the app is being ran either via gradle or via java as a jar.