🧑‍💻 Initial commit on custom Command Framework
This command framework attempts to structure command objects into a
tree structure which can be easily serialized into a JDA slash command
from the root element of the tree.

The idea here was originally to do DDD, however, I accidentally
programmed against JDA interfaces, so this will be remade in a future
commit to comply with DDD principles.