🧑‍💻 Initial commit on the Module System
Due to a possible race condition depending on how Spring decides to
wire my dependencies, and how commands are registered, I decided to
implement a module system.

Each feature will be a Module, which has an internal and public API.
The public API can be used by other modules to talk to each other.
For example, Logging would want to get notified by other modules when
something needs to be logged.

The module class lives in Domain. Because of this, and me not wanting
to duplicate half the code defined in JDA, the event listener list takes
any non-primitive type. As a result of my terrible architectural
decisions, I now need to validate the type of listeners every time I
want to work with the Module... an acceptable sacrifice for progression.